Changes forgot password page to a form and moves directory
